requirement.txt
honcho==1.0.1
ipython==6.0.0
Flask==0.12.1
PyMySQL==0.7.11
python-envcfg-0.2.0
flask-script-2.0.5
flask-bootstrap-3.3.7.1
flask-moment-0.5.1
flask-wtf-0.14.2
flask-sqlalchemy-2.2